refactor: merge capa_base.py into capa_module.py
