+++ openedx/features/personalized_learner_schedules/show_answer/show_answer_field_override.py
@@ -3,7 +3,7 @@ FieldOverride that forces Show Answer values that use Past Due logic to
 new Show Answer values that remove the Past Due check (keeping the rest intact)
 """
 
-from common.lib.xmodule.xmodule.capa_base import SHOWANSWER
+from common.lib.xmodule.xmodule.capa_module import SHOWANSWER
 from lms.djangoapps.courseware.field_overrides import FieldOverrideProvider
 from openedx.features.course_experience import RELATIVE_DATES_FLAG
