+++ openedx/features/personalized_learner_schedules/show_answer/tests/test_show_answer_override.py
@@ -6,7 +6,7 @@ from django.test.utils import override_settings
 
 from edx_toggles.toggles.testutils import override_waffle_flag
 
-from common.lib.xmodule.xmodule.capa_base import SHOWANSWER
+from common.lib.xmodule.xmodule.capa_module import SHOWANSWER
 from lms.djangoapps.ccx.tests.test_overrides import inject_field_overrides
 from lms.djangoapps.courseware.model_data import FieldDataCache
 from lms.djangoapps.courseware.module_render import get_module
